requirements.txt
fastapi
uvicorn
mysql-connector-python
playwright
python-dateutil
statistics
bs4
aiomysql
python-dotenv
